+++ libraries/transformer/src/main/java/androidx/media3/transformer/TransformerUtil.java
@@ -93,6 +93,9 @@ import com.google.common.collect.ImmutableList;
     if (!firstEditedMediaItem.effects.audioProcessors.isEmpty()) {
       return true;
     }
+    if (!composition.effects.audioProcessors.isEmpty()) {
+      return true;
+    }
     return false;
   }